+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 HTTP client extension Change Log
 2.0.8 under development
 -----------------------
 
-- no changes in this release.
+- Bug #168: `Response::detectFormatByContent` falsely detected HTML as XML (CeBe)
 
 
 2.0.7 September 24, 2018