+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Yii Framework 2 HTTP client extension Change Log
 -----------------------
 
 - Enh #153: Allow configuring `JsonParser` to parse JSON as objects instead of arrays (CeBe)
+- Bug #168: `Response::detectFormatByContent` falsely detected HTML as XML (CeBe)
 
 
 2.0.7 September 24, 2018