+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Yii Framework 2 HTTP client extension Change Log
 
 - Enh #153: Allow configuring `JsonParser` to parse JSON as objects instead of arrays (CeBe)
 - Bug #168: `Response::detectFormatByContent` falsely detected HTML as XML (CeBe)
+- Bug #173: Added extra check to `Message::addData()` to prevent error on trying to merge non-array (samdark)
 
 
 2.0.7 September 24, 2018